# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agent Framework API",
        "title": "GoAgent API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/agentos/runs": {
            "post": {
                "description": "Start a generic AgentOS run on the selected backend.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Start AgentOS run",
                "operationId": "agentos-start-run",
                "parameters": [
                    {
                        "description": "AgentOS run request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AgentOSStart"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_control.RunStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/agentos/runs/{run_id}/control": {
            "post": {
                "description": "Send lifecycle control such as pause, resume, or cancel to an AgentOS run.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Control AgentOS run",
                "operationId": "agentos-control-run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "AgentOS control operation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AgentOSControl"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/agentos/runs/{run_id}/signals": {
            "post": {
                "description": "Send business input such as user.message to an AgentOS run.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Signal AgentOS run",
                "operationId": "agentos-signal-run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "AgentOS signal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AgentOSSignal"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/agentos/runs/{run_id}/status": {
            "get": {
                "description": "Query the current status of an AgentOS run.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agentos"
                ],
                "summary": "Get AgentOS run status",
                "operationId": "agentos-run-status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_control.RunStatus"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "control.BackendKind": {
            "type": "string",
            "enum": [
                "native",
                "temporal_external",
                "http",
                "grpc"
            ],
            "x-enum-varnames": [
                "BackendKindNative",
                "BackendKindTemporalExternal",
                "BackendKindHTTP",
                "BackendKindGRPC"
            ]
        },
...
        "control.PlanBudgetUsage": {
            "type": "object",
            "properties": {
                "spent_cents": {
                    "type": "integer"
                }
            }
        },
...
        "core.ArtifactKind": {
            "type": "string",
            "enum": [
                "object",
                "text",
                "file",
                "patch",
                "report",
                "reference",
                "plan_delta"
            ],
            "x-enum-varnames": [
                "ArtifactKindObject",
                "ArtifactKindText",
                "ArtifactKindFile",
                "ArtifactKindPatch",
                "ArtifactKindReport",
                "ArtifactKindReference",
                "ArtifactKindPlanDelta"
            ]
        },
        "core.ArtifactRef": {
            "type": "object",
            "properties": {
                "artifact_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "digest": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/core.ArtifactKind"
                },
                "media_type": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "node_id": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "core.RunProgress": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_TekkenSteve_GoAgent_agentos_control.BackendRef": {
            "type": "object",
            "properties": {
                "kind": {
                    "$ref": "#/definitions/control.BackendKind"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_TekkenSteve_GoAgent_agentos_control.RunStatus": {
            "type": "object",
            "properties": {
                "artifacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ArtifactRef"
                    }
                },
                "budget_usage": {
                    "$ref": "#/definitions/control.PlanBudgetUsage"
                },
                "lifecycle_state": {
                    "type": "string"
                },
                "progress": {
                    "$ref": "#/definitions/core.RunProgress"
                },
                "reason": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_TekkenSteve_GoAgent_agentos_core.ControlOperation": {
            "type": "string",
            "enum": [
                "pause",
                "resume",
                "cancel"
            ],
            "x-enum-varnames": [
                "ControlPause",
                "ControlResume",
                "ControlCancel"
            ]
        },
...
        "github_com_TekkenSteve_GoAgent_agentos_core.SignalType": {
            "type": "string",
            "enum": [
                "control.pause",
                "control.resume",
                "control.cancel",
                "plan.node.retry",
                "plan.approve",
                "plan.reject",
                "user.message",
                "user.approval",
                "user.reject",
                "tool.result",
                "human.feedback",
                "config.patch",
                "memory.patch"
            ],
            "x-enum-varnames": [
                "SignalControlPause",
                "SignalControlResume",
                "SignalControlCancel",
                "SignalPlanNodeRetry",
                "SignalPlanApprove",
                "SignalPlanReject",
                "SignalUserMessage",
                "SignalUserApproval",
                "SignalUserReject",
                "SignalToolResult",
                "SignalHumanFeedback",
                "SignalConfigPatch",
                "SignalMemoryPatch"
            ]
        },
        "request.AgentOSControl": {
            "type": "object",
            "required": [
                "operation"
            ],
            "properties": {
                "actor_id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "operation": {
                    "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_core.ControlOperation"
                },
                "requested_at": {
                    "type": "string"
                }
            }
        },
...
        "request.AgentOSSignal": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "actor_id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "sent_at": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_core.SignalType"
                }
            }
        },
        "request.AgentOSStart": {
            "type": "object",
            "required": [
                "account_id",
                "backend",
                "run_id"
            ],
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "agent_id": {
                    "type": "string"
                },
                "backend": {
                    "$ref": "#/definitions/github_com_TekkenSteve_GoAgent_agentos_control.BackendRef"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "input": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "model_ref": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "requested_at": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "system_prompt": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "string"
                },
                "user_message": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
...
    }
}